requirements.txt
# PDF Processing
PyPDF2>=3.0.0
pdf2image>=1.16.3
pytesseract>=0.3.10
Pillow>=10.0.0
PyMuPDF>=1.23.0

# Text Processing
nltk>=3.8.1
langdetect>=1.0.9
spacy>=3.7.2

# Vector Processing
sentence-transformers>=2.2.2
numpy>=1.24.0
torch>=2.0.0

# Vector Database
chromadb>=0.4.0
pydantic>=2.0.0

# API
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
sqlalchemy==2.0.27
psycopg2-binary==2.9.9
pydantic==2.6.1
pydantic-settings==2.1.0
python-dotenv==1.0.1

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0
httpx>=0.24.0

# Development
black>=23.7.0
isort>=5.12.0
flake8>=6.1.0
mypy>=1.5.0 